renamed activityUtils to activityQuery
