+++ packages/commonwealth/server/routes/viewUserActivity.ts
@@ -1,6 +1,6 @@
 import { AppError } from 'common-common/src/errors';
 import type { NextFunction, Request, Response } from 'express';
-import { getActivityFeed } from '../util/activityUtils';
+import { getActivityFeed } from '../util/activityQuery';
 import type { DB } from '../models';
 
 export const Errors = {
